完成了，可以連結前後端並且評分機制也大致可行，但是等待畫面目前還沒辦法克服

+++ Final/src/main/java/WebTree.java
@@ -26,7 +26,7 @@ public class WebTree{
         nodeList.sort(new Comparator<WebNode>() {
             @Override
             public int compare(WebNode node1, WebNode node2) {
-            	return Integer.compare(node2.nodeScore, node1.nodeScore);
+            	return Double.compare(node2.nodeScore, node1.nodeScore);
             }
         });
